docs : 요구사항 리스트 작성
