+++ README.md
@@ -1,20 +1,42 @@
 
 # 요구사항 리스트
 
-### TODO
+## TODO
 
 
+### PersonResult
+- [ ] 문자열로 이름 입력 시 이름, 결과 Map 반환
+- [ ] 문자열로 all 입력 시 List 이름, 결과 map 반환
+- [ ] 문자열로 없는 이름 입력 시 예외 발생 
+
+### Persons
+- [ ] index번호로 Person 객체 반환
+
+### Results (결과 리스트)
+- [ ] index 번호로 Result 객체 반환
+
+
+### Result (결과 1개)
+- [ ] 빈 값 검증
+
+
+### Ladders (사다리)
+- [ ] Person 객체로 Result 결과 반환
+
+
+### Ladder (사다리 1줄)
+- [ ] index 번호로 가야할 index(index -1 , index, index +1) 번호 반환    
 
 ## DONE
 
-### Heights (높이 리스트)
+### Ladders (사다리)
 - [x] 높이 1미만이면 예외 처리
+
+### Ladder (사다리 1줄)
+- [x] 전략 패턴 검증
 
 ### Persons (참여 인원들)
 - [x] 배열 입력 시 리스트 검증
 
 ### Person (참여 인원)
 - [x] 공백, 최대 길이 검증
-
-### Lines(높이 1줄의 줄 여부)
-- [x] 전략 패턴 검증
