+++ src/main/java/nextstep/ladder/controller/LadderController.java
@@ -1,6 +1,6 @@
 package nextstep.ladder.controller;
 
-import nextstep.ladder.domain.Ladder;
+import nextstep.ladder.domain.Ladders;
 import nextstep.ladder.domain.Persons;
 import nextstep.ladder.view.InputView;
 import nextstep.ladder.view.ResultView;
@@ -11,7 +11,7 @@ public class LadderController {
         Persons persons = new Persons(InputView.requestNames());
         int heightCount = InputView.requestHeightCount();
 
-        ResultView.print(persons, new Ladder(persons.personCount(), heightCount));
+        ResultView.print(persons, new Ladders(persons.personCount(), heightCount));
     }
 
 }
